Changes in the interface
The quotes now are aligned to the left.
The text for the user on the terminal was changed.
The time for the command's animation was reduced and its text was changed.
The status bar color was changed.
The background was rotated to portrait mode.
The app were locked to be seen only in portrait mode.
The name of the app was changed.